vorbis: Residue decoding performance improvements and clean-up.
If no codebooks are assigned to a residue pass, do not perform the
pass since it would do nothing but waste CPU cycles.

This provides a modest measurable performance increase.

+++ symphonia-codec-vorbis/src/common.rs
@@ -17,7 +17,7 @@ pub fn ilog(x: u32) -> u32 {
 pub struct BitSetIterator<'a> {
     bits: &'a [u32],
     pos: usize,
-    count: u32,
+    count: usize,
 }
 
 impl<'a> Iterator for BitSetIterator<'a> {
@@ -53,7 +53,7 @@ macro_rules! decl_bitset {
         #[derive(Default)]
         pub struct $name {
             bits: [u32; $size >> 5],
-            bit_count: u32,
+            bit_count: usize,
         }
 
         impl $name {
@@ -79,7 +79,7 @@ macro_rules! decl_bitset {
             }
 
             #[inline(always)]
-            pub fn count(&self) -> u32 {
+            pub fn count(&self) -> usize {
                 self.bit_count
             }
